+++ packages/dataviews/src/field-types/index.tsx
@@ -22,6 +22,7 @@ import { default as boolean } from './boolean';
 import { default as media } from './media';
 import { default as array } from './array';
 import { default as telephone } from './telephone';
+import { default as url } from './url';
 import { renderFromElements } from '../utils';
 import { ALL_OPERATORS, OPERATOR_IS, OPERATOR_IS_NOT } from '../constants';
 
@@ -68,6 +69,10 @@ export default function getFieldTypeDefinition< Item >(
 
 	if ( 'telephone' === type ) {
 		return telephone;
+	}
+
+	if ( 'url' === type ) {
+		return url;
 	}
 
 	// This is a fallback for fields that don't provide a type.